+++ src/main/java/com/google/firebase/iid/FirebaseInstanceId.java
@@ -79,7 +79,7 @@ public class FirebaseInstanceId {
     checkArgument(!Strings.isNullOrEmpty(projectId),
         "Project ID is required to access instance ID service. Use a service account credential or "
             + "set the project ID explicitly via FirebaseOptions. Alternatively you can also "
-            + "set the project ID via the GCLOUD_PROJECT environment variable.");
+            + "set the project ID via the GOOGLE_CLOUD_PROJECT environment variable.");
   }
 
   /**